// src/gql/graphql.ts
export type ProductListItemFragment = { id?: string | null, attributes?: { name: string, description: string, price: number, slug: string, images: { data: Array<{ attributes?: { url: string, alternativeText?: string | null } | null }> }, categories?: { data: Array<{ id?: string | null, attributes?: { name: string, slug: string } | null }> } | null } | null };